Refactored the code to fit MVC pattern.
