+++ Application/src/controller/Import.java
@@ -1,4 +1,4 @@
-package model;
+package controller;
 
 import javafx.scene.control.Alert;
 import javafx.stage.FileChooser;
