+++ Application/src/controller/network/Net.java
@@ -1,6 +1,6 @@
-package net;
+package controller.network;
 
-import visuals.ExecutionLog;
+import view.ExecutionLog;
 import javafx.application.Platform;
 
 import java.util.PriorityQueue;
@@ -27,7 +27,7 @@ public class Net implements Runnable {
     }
 
     /**
-     * Only allow a single network thread and it must be run on an FXApplicationThread.
+     * Only allow a single controller.network thread and it must be run on an FXApplicationThread.
      * Starts a new Net thread.
      */
     public static void init() {
