+++ Application/src/view/visuals/Renderable.java
@@ -1,6 +1,7 @@
-package visuals;
+package view.visuals;
 
 import javafx.scene.canvas.GraphicsContext;
+import view.visuals.component.Coordinates;
 
 /**
  * Implementing this assures the class can be rendered through generics.
